backend/requirements.txt
fastapi
uvicorn
pydantic
faststream[redis]
aiofiles
redis
google-generativeai